properly insert variant versions for x.x in outputs
